+++ src/test/scala/vectorpipe/vectortile/LayerTestPipeline.scala
@@ -1,15 +1,16 @@
 package vectorpipe.vectortile
 
+import java.net.URI
+
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions
 import org.apache.spark.sql.functions.when
 import org.locationtech.jts.{geom => jts}
-
 import vectorpipe._
 import vectorpipe.functions.osm._
 import vectorpipe.vectortile._
 
-case class LayerTestPipeline(geometryColumn: String, baseOutputURI: java.net.URI) extends Pipeline {
+case class LayerTestPipeline(geometryColumn: String, baseOutputURI: Option[URI]) extends Pipeline {
   val layerMultiplicity = LayerNamesInColumn("layers")
 
   override def select(wayGeoms: DataFrame, targetZoom: Int, keyColumn: String): DataFrame = {