+++ src/test/scala/vectorpipe/vectortile/PipelineSpec.scala
@@ -1,5 +1,7 @@
 package vectorpipe.vectortile
 
+import java.net.URI
+
 import org.apache.spark.sql.functions
 import org.apache.spark.sql.functions.{isnull, lit}
 import org.locationtech.geomesa.spark.jts._
@@ -26,17 +28,17 @@ class PipelineSpec extends FunSpec with TestEnvironment with Matchers {
     val wayGeoms = vp.reconstructWayGeometries(df, nodes).cache
 
     it("should generate a single zoom level") {
-      val pipeline = TestPipeline("geometry", new java.net.URI("file:///tmp/iom-tiles"), 16)
+      val pipeline = TestPipeline("geometry", Some(new URI("file:///tmp/iom-tiles")), 16)
       VectorPipe(nodeGeoms, pipeline, VectorPipe.Options.forZoom(8))
     }
 
     it("should generate multiple zoom levels") {
-      val pipeline = TestPipeline("geometry", new java.net.URI("file:///tmp/iom-tiles-pyramid"), 16)
+      val pipeline = TestPipeline("geometry", Some(new URI("file:///tmp/iom-tiles-pyramid")), 16)
       VectorPipe(nodeGeoms, pipeline, VectorPipe.Options.forZoomRange(6, 8))
     }
 
     it("should generate multiple layers") {
-      val pipeline = LayerTestPipeline("geom", new java.net.URI("file:///tmp/iom-layers"))
+      val pipeline = LayerTestPipeline("geom", Some(new URI("file:///tmp/iom-layers")))
       VectorPipe(wayGeoms, pipeline, VectorPipe.Options.forZoom(14))
     }
   }